LG4 Rename module scopes_conditions to scope_conditions
