+++ model/scope_conditions.py
